Reject Workload API responses missing the bundle
The Workload API specification says that clients SHOULD reject messages
that with unset fields in the SVID message. The bundle is part of that
message. The current code allows an empty bundle field. In particular,
this means that the FetchX509Context and FetchX509Bundles calls can
return empty bundle contents.

Considering the "SHOULD" guidance, this change updates the client to
reject empty bundles. This only impacts the X.509 related calls that
return bundle information, but not calls that only return the SVID (e.g.
FetchX509SVID).

I'd be open to being more restrictive and downright rejecting messages
from the Workload API that have missing fields in the SVID message of
any kind, but this seemed like a good minimal step.

Also took the opportunity to DRY up some of the response handling code.

Signed-off-by: Andrew Harding <andrew.harding@hpe.com>